moved auth into navbar
